requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
pinecone-client==3.1.0
openai==1.12.0
huggingface_hub==0.20.3
pypdf==4.0.1
python-dotenv==1.0.1
requests==2.31.0
httpx==0.26.0
